+++ src/enums/generators/identifier-names-generators/IdentifierNamesGenerator.ts
@@ -1,5 +1,11 @@
-export enum IdentifierNamesGenerator {
-    DictionaryIdentifierNamesGenerator = 'dictionary',
-    HexadecimalIdentifierNamesGenerator = 'hexadecimal',
-    MangledIdentifierNamesGenerator = 'mangled'
-}
+import { MakeEnum } from '@gradecam/tsenum';
+
+export const IdentifierNamesGenerator: Readonly<{
+    MangledIdentifierNamesGenerator: string;
+    DictionaryIdentifierNamesGenerator: string;
+    HexadecimalIdentifierNamesGenerator: string;
+}> = MakeEnum({
+    DictionaryIdentifierNamesGenerator: 'dictionary',
+    HexadecimalIdentifierNamesGenerator: 'hexadecimal',
+    MangledIdentifierNamesGenerator: 'mangled'
+});